requirements.txt
ahk-binary==2023.9.0
altgraph==0.17.4
contourpy==1.3.1
cycler==0.12.1
fire==0.5.0
fonttools==4.55.3
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
opencv-python==4.7.0.72
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pybboxes==0.1.6
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMsgBox==1.0.9
pyparsing==3.2.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pyqtgraph==0.13.7
pyserial==3.5
pyserial-asyncio==0.6
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
scipy==1.14.1
shapely==2.0.2
six==1.17.0
tensorflow==2.18.0
terminaltables==3.1.10
